Enhance HashTable with Object.observe to handle dynamically changing keys
This feature is untestable in the current node version since it doesn't
use the latest V8 build. But it is possible to test in latest Chrome Canary
build with experimental JavaScript features turned on in about:flags.

More info on Object.observe:
http://wiki.ecmascript.org/doku.php?id=harmony:observe

Use the following HTML to test in Chrome Canary:

-------------------------------------------------------------------------------

<html>
<head>
  <script type="text/javascript" src="./lib/structure.js"></script>
</head>
<body>
<script type="text/javascript">
  var ht = new structure.HashTable();
  var key = {c : "4"};
  var value = {a : 5};
  ht.put(key, value);
  key.d = 324;
  console.log(ht.get(key));  // should return value
  console.log(ht.get({c : "4"}));  // should not return value
  console.log(ht.get({c : "4", d: 324}));  // should return value
</script>
</body>
</html>